==> research/03_data_trasformation.ipynb <==
import os
%pwd

os.chdir("../")
%pwd

import pandas as pd
data=pd.read_csv("artifacts/data_ingestion/winequality-red.csv")
data.head()

data.info()


from  dataclasses  import dataclass
from  pathlib  import Path

@dataclass
class DataTransformationConfig:
    root_dir: Path
    data_path: Path

from src.datascience.constants import *
from src.datascience.utils.common import read_yaml,create_directories

class ConfigurationManager:
    def __init__(self,
                 config_filepath=CONFIG_FILE_PATH,
                 params_filepath = PARAMS_FILE_PATH,
                 schema_filepath = SCHEMA_FILE_PATH):
        self.config=read_yaml(config_filepath)
        self.params=read_yaml(params_filepath)
        self.schema=read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_transformation_config(self)->DataTransformationConfig:
        config=self.config.data_transformation
        create_directories([config.root_dir])
        data_transformation_config=DataTransformationConfig(
            root_dir=Path(config.root_dir),
            data_path=Path(config.data_path)
        )
        return data_transformation_config



import os
from src.datascience import logger
from sklearn.model_selection import train_test_split
import pandas as pd


class DataTransformation:
    def __init__(self, config):
        self.config = config
        
    
    def train_test_splitting(self):
        data=pd.read_csv(self.config.data_path)
        train, test = train_test_split(data, test_size=0.2, random_state=42)
        train.to_csv(os.path.join(self.config.root_dir, "train.csv"), index=False)
        test.to_csv(os.path.join(self.config.root_dir, "test.csv"), index=False)
        logger.info("Train and test data split and saved successfully.")
        logger.info(f"Train data shape: {train.shape}, Test data shape: {test.shape}")

        print(train.shape)
        print(test.shape)

try:
    config_manager = ConfigurationManager()
    data_transformation_config = config_manager.get_data_transformation_config()
    data_transformation = DataTransformation(data_transformation_config)
    data_transformation.train_test_splitting()
except Exception as e:
    raise e
